% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_equity.R
\name{dp_equity}
\alias{dp_equity}
\title{Compare path descriptors across groups}
\usage{
dp_equity(x, group)
}
\arguments{
\item{x}{A decision_path object}

\item{group}{Grouping variable name as a character string}
}
\value{
A tibble of grouped summaries
}
\description{
Produces simple subgroup summaries for key decision-path descriptors.
}
