% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.dp_entropy}
\alias{plot.dp_entropy}
\title{Plot a dp_entropy object}
\usage{
\method{plot}{dp_entropy}(x, top = 10L, ...)
}
\arguments{
\item{x}{A \code{dp_entropy} object from \code{\link{dp_entropy}}.}

\item{top}{Integer. Number of top paths to display. Default 10.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot2} object.
}
\description{
Produces a bar chart of the most frequent decision paths.
}
