% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisionpaths-package.R
\docType{package}
\name{decisionpaths-package}
\alias{decisionpaths}
\alias{decisionpaths-package}
\title{decisionpaths: Construct and audit longitudinal decision paths}
\description{
Tools for constructing and auditing decision paths from panel data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/subirhait/decisionpaths}
  \item Report bugs at \url{https://github.com/subirhait/decisionpaths/issues}
}

}
\author{
\strong{Maintainer}: Subir Hait \email{haitsubi@msu.edu} (\href{https://orcid.org/0009-0004-9871-9677}{ORCID})

}
\keyword{internal}
